feat: move banner to hook
